build: run perf test queries only on specific datasets

+++ scripts/ci/run_perftest.sh
@@ -166,48 +166,30 @@ force_compaction() {
 # must be provided to both the data generation and query generation commands
 # and must be the same to ensure that the queries cover the data range.
 start_time() {
-  case $1 in
-    iot|window-agg|group-agg|bare-agg)
-      echo 2018-01-01T00:00:00Z
-      ;;
-    group-window-transpose)
-      cardinality=$(echo $2 | cut -d '-' -f2)
-      if [ "$cardinality" = "low" ]; then
-        echo 2018-01-01T00:00:00Z
-      else
-        echo 2019-01-01T00:00:00Z
-      fi
-      ;;
-    metaquery)
-      echo 2019-01-01T00:00:00Z
-      ;;
-    multi-measurement)
-      echo 2017-01-01T00:00:00Z
-      ;;
-    *)
-      echo "unknown use-case: $1"
-      exit 1
-      ;;
-  esac
+  # All queries and datasets can start at the same time. Certain queries and
+  # datasets will use case-dependent end-times.
+  echo 2018-01-01T00:00:00Z
 }
+
 end_time() {
   case $1 in
     iot|window-agg|group-agg|bare-agg)
       echo 2018-01-01T12:00:00Z
       ;;
+    multi-measurement|metaquery)
+      echo 2019-01-01T00:00:00Z
+      ;;
     group-window-transpose)
+    # The group-window transpose tests currently require special handling due to
+    # the difference in performance between having the existing pushdown enabled
+    # or disabled for data with varying degrees of cardinality. Ideally this
+    # pushdown will be optimized and this test can be simplified in the future.
       cardinality=$(echo $2 | cut -d '-' -f2)
       if [ "$cardinality" = "low" ]; then
         echo 2018-01-01T12:00:00Z
       else
-        echo 2020-01-01T00:00:00Z
+        echo 2019-01-01T00:00:00Z
       fi
-      ;;
-    metaquery)
-      echo 2020-01-01T00:00:00Z
-      ;;
-    multi-measurement)
-      echo 2018-01-01T00:00:00Z
       ;;
     *)
       echo "unknown use-case: $1"
@@ -215,37 +197,6 @@ end_time() {
       ;;
   esac
 }
-
-# Run and record tests
-
-# Generate and ingest bulk data. Record the time spent as an ingest test.
-for usecase in iot metaquery multi-measurement; do
-  data_fname="influx-bulk-records-usecase-$usecase"
-  $GOPATH/bin/bulk_data_gen \
-      -seed=$seed \
-      -use-case=$usecase \
-      -scale-var=$scale_var \
-      -timestamp-start=$(start_time $usecase) \
-      -timestamp-end=$(end_time $usecase) > \
-    ${DATASET_DIR}/$data_fname
-
-  load_opts="-file=${DATASET_DIR}/$data_fname -batch-size=$batch -workers=$workers -urls=http://${NGINX_HOST}:8086 -do-abort-on-exist=false -do-db-create=true -backoff=1s -backoff-timeout=300m0s"
-  if [ -z $INFLUXDB2 ] || [ $INFLUXDB2 = true ]; then
-    load_opts="$load_opts -organization=$TEST_ORG -token=$TEST_TOKEN"
-  fi
-
-  # Run ingest tests. Only write the results to disk if this run should contribute to ingest-test results.
-  out=/dev/null
-  if [ "${TEST_RECORD_INGEST_RESULTS}" = true ]; then
-    out=$working_dir/test-ingest-$usecase.json
-  fi
-  $GOPATH/bin/bulk_load_influx $load_opts | \
-    jq ". += {branch: \"$INFLUXDB_VERSION\", commit: \"$TEST_COMMIT\", time: \"$datestring\", i_type: \"$DATA_I_TYPE\", use_case: \"$usecase\"}" > ${out}
-
-  # Cleanup
-  force_compaction
-  rm ${DATASET_DIR}/$data_fname
-done
 
 query_types() {
   case $1 in
@@ -271,6 +222,31 @@ query_types() {
   esac
 }
 
+# Many of the query generator use-cases have aliases to make reporting more
+# clear. This function will translate the aliased query use cases to their
+# dataset use cases.
+query_usecase_alias() {
+  case $1 in
+    window-agg|group-agg|bare-agg|group-window-transpose|iot)
+      echo iot
+      ;;
+    metaquery)
+      echo metaquery
+      ;;
+    multi-measurement)
+      echo multi-measurement
+      ;;
+    *)
+      echo "unknown use-case: $1"
+      exit 1
+      ;;
+  esac
+}
+
+##########################
+## Run and record tests ##
+##########################
+
 # Generate queries to test.
 query_files=""
 for usecase in window-agg group-agg bare-agg group-window-transpose iot metaquery multi-measurement; do
@@ -289,28 +265,69 @@ for usecase in window-agg group-agg bare-agg group-window-transpose iot metaquer
   done
 done
 
-# Run the query benchmarks
-for query_file in $query_files; do
-  format=$(echo $query_file | cut -d '_' -f1)
-  usecase=$(echo $query_file | cut -d '_' -f2)
-  type=$(echo $query_file | cut -d '_' -f3)
-  ${GOPATH}/bin/query_benchmarker_influxdb \
-      -file=${DATASET_DIR}/$query_file \
-      -urls=http://${NGINX_HOST}:8086 \
-      -debug=0 \
-      -print-interval=0 \
-      -json=true \
-      -workers=$workers \
-      -benchmark-duration=$duration | \
-    jq '."all queries"' | \
-    jq -s '.[-1]' | \
-    jq ". += {use_case: \"$usecase\", query_type: \"$type\", branch: \"$INFLUXDB_VERSION\", commit: \"$TEST_COMMIT\", time: \"$datestring\", i_type: \"$DATA_I_TYPE\", query_format: \"$format\"}" > \
-      $working_dir/test-query-$format-$usecase-$type.json
+# Generate and ingest bulk data. Record the time spent as an ingest test if
+# specified, and run the query performance tests for each dataset.
+for usecase in iot metaquery multi-measurement; do
+  data_fname="influx-bulk-records-usecase-$usecase"
+  $GOPATH/bin/bulk_data_gen \
+      -seed=$seed \
+      -use-case=$usecase \
+      -scale-var=$scale_var \
+      -timestamp-start=$(start_time $usecase) \
+      -timestamp-end=$(end_time $usecase) > \
+    ${DATASET_DIR}/$data_fname
 
-    # restart daemon
-    systemctl stop influxdb
-    systemctl unmask influxdb.service
-    systemctl start influxdb
+  load_opts="-file=${DATASET_DIR}/$data_fname -batch-size=$batch -workers=$workers -urls=http://${NGINX_HOST}:8086 -do-abort-on-exist=false -do-db-create=true -backoff=1s -backoff-timeout=300m0s"
+  if [ -z $INFLUXDB2 ] || [ $INFLUXDB2 = true ]; then
+    load_opts="$load_opts -organization=$TEST_ORG -token=$TEST_TOKEN"
+  fi
+
+  # Run ingest tests. Only write the results to disk if this run should contribute to ingest-test results.
+  out=/dev/null
+  if [ "${TEST_RECORD_INGEST_RESULTS}" = true ]; then
+    out=$working_dir/test-ingest-$usecase.json
+  fi
+  $GOPATH/bin/bulk_load_influx $load_opts | \
+    jq ". += {branch: \"$INFLUXDB_VERSION\", commit: \"$TEST_COMMIT\", time: \"$datestring\", i_type: \"$DATA_I_TYPE\", use_case: \"$usecase\"}" > ${out}
+
+  # Cleanup from the data generation and loading.
+  force_compaction
+  rm ${DATASET_DIR}/$data_fname
+ 
+  # Run the query tests applicable to this dataset.
+  for query_file in $query_files; do 
+    format=$(echo $query_file | cut -d '_' -f1)
+    query_usecase=$(echo $query_file | cut -d '_' -f2)
+    type=$(echo $query_file | cut -d '_' -f3)
+
+    # Only run the query tests for queries applicable to this dataset.
+    if [ "$usecase" != "$(query_usecase_alias $query_usecase)" ]; then
+      continue
+    fi
+
+    ${GOPATH}/bin/query_benchmarker_influxdb \
+        -file=${DATASET_DIR}/$query_file \
+        -urls=http://${NGINX_HOST}:8086 \
+        -debug=0 \
+        -print-interval=0 \
+        -json=true \
+        $(org_flag $format) \
+        -token=$TEST_TOKEN \
+        -workers=$workers \
+        -benchmark-duration=$duration | \
+      jq '."all queries"' | \
+      jq -s '.[-1]' | \
+      jq ". += {use_case: \"$query_usecase\", query_type: \"$type\", branch: \"$INFLUXDB_VERSION\", commit: \"$TEST_COMMIT\", time: \"$datestring\", i_type: \"$DATA_I_TYPE\", query_format: \"$format\"}" > \
+        $working_dir/test-query-$format-$query_usecase-$type.json
+
+      # Restart daemon between query tests.
+      systemctl stop influxdb
+      systemctl unmask influxdb.service
+      systemctl start influxdb
+  done
+
+  # Delete DB to start anew.
+  influx -database $db_name -execute "DROP DATABASE $db_name"
 done
 
 echo "Using Telegraph to report results from the following files:"
